fix(qr): #99 handle file drop globally in useDragAndDrop and SidePanelLayout to resolve empty space drop issue
